# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "APIs",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a new user.",
        "description": "User registration.",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "User details for registration.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddNewUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Access token and refresh token for the newly registered user.",
            "schema": {
              "type": "object",
              "properties": {
                "refreshToken": {
                  "$ref": "#/definitions/Token"
                },
                "accessToken": {
                  "$ref": "#/definitions/Token"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "User with the same email already exists."
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login to the application",
        "description": "Logs in to an existing user and returns access and refresh tokens.",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "Write there your Username",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "someUsername"
                },
                "password": {
                  "type": "string",
                  "example": "somePassword"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns access token and refresh token.",
            "schema": {
              "type": "object",
              "properties": {
                "refreshToken": {
                  "$ref": "#/definitions/Token"
                },
                "accessToken": {
                  "$ref": "#/definitions/Token"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad request. Username or password missing."
          },
          "403": {
            "description": "Forbidden. Invalid login credentials. Password doesnt match User."
          }
        }
      }
    },
    "/api/auth/refreshToken": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Refresh user",
        "description": "Get a new access token by sending a valid refresh token",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "refreshToken": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns access token and refresh token.",
            "schema": {
              "type": "object",
              "properties": {
                "refreshToken": {
                  "$ref": "#/definitions/Token"
                },
                "accessToken": {
                  "$ref": "#/definitions/Token"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Missing refresh token"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Verification of refresh token failed"
          }
        }
      }
    },
    "/api/auth/revokeRefreshTokens": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Revoke refresh tokens for a user",
        "description": "Revokes all the refresh tokens of a user by their userId",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AddNewUser": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "example": "SomeName"
        },
        "profileName": {
          "type": "string",
          "example": "Dude 1"
        },
        "email": {
          "type": "string",
          "example": "thisIsTheemail@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "SomeGoodpassword123"
        }
      },
      "required": [
        "username",
        "profileName",
        "email",
        "password"
      ]
    },
...
    "Token": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        }
      }
    }
  }
}